fix : 유저 프로필 컴포넌트화
